+ Qt version update
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfo/trunk@3335 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/GUI/Qt/sheetview.cpp
@@ -34,7 +34,7 @@ SheetView::~SheetView()
 }
 
 void SheetView::refreshDisplay() {
-    ui->tableWidget->setRowCount(C->Count_Get());
+    ui->tableWidget->setRowCount((int)C->Count_Get());
     ui->tableWidget->setColumnCount(Sheet::getSheet()->getNbColumns());
     for(int i=0;i<Sheet::getSheet()->getNbColumns();++i) {
         column c = Sheet::getSheet()->getColumn(i);
